Moved _dataset to utils; default Dataset has transform, target_transform and transforms and EventSample type.
